+++ internal/provider/provider.go
@@ -124,6 +124,7 @@ func (p *TerraProbeProvider) Resources(ctx context.Context) []func() resource.Re
 	return []func() resource.Resource{
 		NewHttpTestResource,
 		NewTcpTestResource,
+		NewTestSuiteResource,
 	}
 }